Allow user to print calibration parameters

+++ bme280_defs.hpp
@@ -221,7 +221,7 @@ public:
         std::cout<<"P paramaters :"<<std::endl;
         std::cout<<"[P1] : "<<m_P1<<" ; "<<"[P2] : "<<m_P2<<" ; "<<"[P3] : "<<m_P3<<"[P4] : "<<m_P4<<" ; "<<"[P5] : "<<m_P5<<" ; "<<"[P6] : "<<m_P6<<"[P7] : "<<m_P7<<" ; "<<"[P8] : "<<m_P8<<" ; "<<"[P9] : "<<m_P9<<" ; "<<std::endl;
         std::cout<<"H paramaters :"<<std::endl;
-        std::cout<<"[H1] : "<<m_H1<<" ; "<<"[H2] : "<<m_H2<<" ; "<<"[H3] : "<<m_H3<<"[H4] : "<<m_H4<<" ; "<<"[H5] : "<<m_H5<<" ; "<<"[H6] : "<<m_H6<<std::endl;
+        std::cout<<"[H1] : "<<(uint8_t)m_H1<<" ; "<<"[H2] : "<<m_H2<<" ; "<<"[H3] : "<<(uint8_t) m_H3<<"[H4] : "<<m_H4<<" ; "<<"[H5] : "<<m_H5<<" ; "<<"[H6] : "<<(int8_t)m_H6<<std::endl;
     }
 	uint16_t m_T1;
 	int16_t m_T2;
